+++ tests/Sulu/Component/Media/SystemCollections/SystemCollectionManagerTest.php
@@ -13,8 +13,8 @@ namespace Sulu\Component\Media\SystemCollections;
 use Doctrine\ORM\EntityManagerInterface;
 use Sulu\Bundle\MediaBundle\Api\Collection;
 use Sulu\Bundle\MediaBundle\Collection\Manager\CollectionManagerInterface;
+use Sulu\Component\Cache\CacheInterface;
 use Sulu\Component\Security\Authentication\UserInterface;
-use Symfony\Component\Config\ConfigCache;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 
@@ -337,17 +337,14 @@ class SystemCollectionManagerTest extends \PHPUnit_Framework_TestCase
      */
     public function testWarmUpNotFresh($config, $existingCollections, $notExistingCollections, $data)
     {
-        $tempName = tempnam('/tmp', 'system_collections');
-        file_put_contents($tempName, serialize($data));
-
         $tokenStorage = $this->getTokenStorage();
         $collectionManager = $this->getCollectionManager($existingCollections, $notExistingCollections);
 
         $entityManager = $this->prophesize(EntityManagerInterface::class);
-        $cache = $this->prophesize(ConfigCache::class);
+        $cache = $this->prophesize(CacheInterface::class);
         $cache->isFresh()->shouldBeCalled()->willReturn(false);
-        $cache->write(serialize($data))->shouldBeCalled();
-        $cache->__toString()->shouldBeCalled()->willReturn($tempName);
+        $cache->write($data)->shouldBeCalled();
+        $cache->read()->shouldBeCalled()->willReturn($data);
 
         $manager = new SystemCollectionManager(
             $config,
@@ -366,17 +363,14 @@ class SystemCollectionManagerTest extends \PHPUnit_Framework_TestCase
      */
     public function testWarmUpFresh($config, $existingCollections, $notExistingCollections, $data)
     {
-        $tempName = tempnam('/tmp', 'system_collections');
-        file_put_contents($tempName, serialize($data));
-
         $tokenStorage = $this->getTokenStorage();
         $collectionManager = $this->getCollectionManager($existingCollections, $notExistingCollections, false);
 
         $entityManager = $this->prophesize(EntityManagerInterface::class);
-        $cache = $this->prophesize(ConfigCache::class);
+        $cache = $this->prophesize(CacheInterface::class);
         $cache->isFresh()->shouldBeCalled()->willReturn(true);
-        $cache->write(serialize($data))->shouldNotBeCalled();
-        $cache->__toString()->shouldBeCalled()->willReturn($tempName);
+        $cache->write($data)->shouldNotBeCalled();
+        $cache->read()->shouldBeCalled()->willReturn($data);
 
         $manager = new SystemCollectionManager(
             $config,
@@ -395,9 +389,6 @@ class SystemCollectionManagerTest extends \PHPUnit_Framework_TestCase
      */
     public function testGetSystemCollection($data, $key, $expected, $exception = null)
     {
-        $tempName = tempnam('/tmp', 'system_collections');
-        file_put_contents($tempName, serialize($data));
-
         if ($exception !== null) {
             $this->setExpectedException($exception);
         }
@@ -406,9 +397,9 @@ class SystemCollectionManagerTest extends \PHPUnit_Framework_TestCase
         $collectionManager = $this->getCollectionManager();
 
         $entityManager = $this->prophesize(EntityManagerInterface::class);
-        $cache = $this->prophesize(ConfigCache::class);
+        $cache = $this->prophesize(CacheInterface::class);
         $cache->isFresh()->shouldBeCalled()->willReturn(true);
-        $cache->__toString()->shouldBeCalled()->willReturn($tempName);
+        $cache->read()->shouldBeCalled()->willReturn($data);
 
         $manager = new SystemCollectionManager(
             [],
@@ -430,16 +421,13 @@ class SystemCollectionManagerTest extends \PHPUnit_Framework_TestCase
      */
     public function testIsSystemCollection($data, $id, $expected)
     {
-        $tempName = tempnam('/tmp', 'system_collections');
-        file_put_contents($tempName, serialize($data));
-
         $tokenStorage = $this->getTokenStorage();
         $collectionManager = $this->getCollectionManager();
 
         $entityManager = $this->prophesize(EntityManagerInterface::class);
-        $cache = $this->prophesize(ConfigCache::class);
+        $cache = $this->prophesize(CacheInterface::class);
         $cache->isFresh()->shouldBeCalled()->willReturn(true);
-        $cache->__toString()->shouldBeCalled()->willReturn($tempName);
+        $cache->read()->shouldBeCalled()->willReturn($data);
 
         $manager = new SystemCollectionManager(
             [],